Refactor to `Bootable` Concern
